Moving sample code to seperate package to being Go work
