+++ go/src/main/java/tech/notpaper/sample/repository/NoteRepository.java
@@ -1,9 +1,9 @@
-package tech.notpaper.go.repository;
+package tech.notpaper.sample.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import tech.notpaper.go.model.Note;
+import tech.notpaper.sample.model.Note;
 
 @Repository
 public interface NoteRepository extends JpaRepository<Note, Long> {
